requirements.txt
requests==2.24.0
torch==1.7.1
numpy==1.18.5
pandas==1.1.2
transformers==4.1.1
glob2==0.7
seaborn==0.11.0
matplotlib==3.3.1
streamlit==0.76.0
torchvision==0.8.2
Pillow==8.1.0
scikit_learn==0.24.1
